# backend/openapi.yaml
openapi: 3.0.3
info:
  title: BookStop API
  description: Highly normalized, ROA-compliant Library Management System API utilizing payload-driven RBAC filtering.
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Book:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string, maxLength: 500 }
        isbn: { type: string, maxLength: 20 }
        stock_quantity: { type: integer }
        true_available_stock: { type: integer, readOnly: true }
        is_archived: { type: boolean, default: false }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        full_name: { type: string }
        role: { type: string, enum: [member, librarian, admin] }
        is_active: { type: boolean, default: true }
    Transaction:
      type: object
      properties:
        id: { type: string, format: uuid }
        book_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        checkout_date: { type: string, format: date-time }
        due_date: { type: string, format: date-time }
        status: { type: string, enum: [issued, returned, overdue, lost] }
    Fine:
      type: object
      properties:
        id: { type: string, format: uuid }
        transaction_id: { type: string, format: uuid }
        amount: { type: number, format: float }
        is_paid: { type: boolean, default: false }
    Reservation:
      type: object
      properties:
        id: { type: string, format: uuid }
        book_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        status: { type: string, enum: [active, fulfilled, cancelled] }
    Notification:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        message: { type: string }
        is_read: { type: boolean, default: false }

security:
  - bearerAuth: []

paths:
  /auth/register:
    post:
      summary: Register Account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201': { description: Created }

  /auth/login:
    post:
      summary: Authenticate Session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200': { description: JWT Access Token returned }

  /book:
    get:
      summary: Catalog Search
      description: Returns all active inventory.
      responses:
        '200':
          description: Array of books.
          content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Book' } } } }
    post:
      summary: Add Asset
      description: Requires admin JWT payload.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/Book' } } }
      responses:
        '201': { description: Created }

  /book/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    
    put:
      summary: Update Asset
      description: Requires admin JWT payload.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/Book' } } }
      responses:
        '200': { description: Updated }
    delete:
      summary: Destroy Asset
      description: Requires admin JWT payload. Enforces FK constraints.
      responses:
        '204': { description: Deleted }

  /user:
    get:
      summary: List Users
      description: Controller inspects JWT. Admin payload -> returns all users. Member payload -> returns array containing only own profile.
      responses:
        '200':
          description: Array of users.
          content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/User' } } } }

  /user/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    
    put:
      summary: Update User
      description: Update profile fields (e.g., full_name, password). Use DELETE /user/{id} for soft delete (deactivation).
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/User' } } }
      responses:
        '200': { description: Updated }
    delete:
      summary: Destroy User
      description: Fails if constraints exist (fines, books).
      responses:
        '204': { description: Deleted }

  /transaction:
    get:
      summary: List Circulation History
      description: Controller inspects JWT. Admin -> global history. Member -> personal history.
      responses:
        '200':
          description: Array of transactions.
          content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Transaction' } } } }
    post:
      summary: Issue Asset (Check-out)
      description: Creates a transaction record. The user_id is forcibly overridden by the JWT sub claim in the middleware to prevent impersonation.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/Transaction' } } }
      responses:
        '201': { description: Issued }

  /transaction/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    
    put:
      summary: Update State (Check-in/Lost)
      description: Controller executes ACID block to free inventory stock and calculate potential fine triggers.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/Transaction' } } }
      responses:
        '200': { description: State updated }
    

  /fine:
    get:
      summary: List Ledger
      description: Controller inspects JWT for role-based filtering.
      responses:
        '200':
          description: Array of fines.
          content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Fine' } } } }
   

  /fine/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    
    put:
      summary: Settle Debt
      description: Set is_paid to true.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/Fine' } } }
      responses:
        '200': { description: Updated }
    

  /reservation:
    get:
      summary: List Holds
      description: Controller inspects JWT for role-based scoping.
      responses:
        '200':
          description: Array of holds.
          content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Reservation' } } } }
    post:
      summary: Create Hold
      description: Enforces DB queue limits. user_id is injected from JWT.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/Reservation' } } }
      responses:
        '201': { description: Created }

  /reservation/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    
    put:
      summary: Resolve Hold
      description: Set status to fulfilled. Use DELETE /reservation/{id} for soft-delete cancellation.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/Reservation' } } }
      responses:
        '200': { description: Updated }
    delete:
      summary: Drop Hold (Admin)
      description: Soft delete. Marks reservation status as cancelled without removing the record.
      responses:
        '204': { description: Deleted }

  /notification:
    get:
      summary: List Alerts
      description: Strict JWT user_id scoping.
      responses:
        '200':
          description: Array of notifications.
          content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Notification' } } } }

  /report:
    get:
      summary: System Analytics (Admin)
      description: Aggregates total inventory, active fines, etc.
      responses:
        '200': { description: Analytics payload }